+++ robot/rover/science/include/Fan.h
@@ -11,7 +11,7 @@
 class Fan :  public Stoppable, public Updatable {
 
     protected:
-        int fanPWM;
+        uint8_t fanPWM;
     public:
         void turnOn();
         void turnOff();
